+++ Hardware/Print/PrintEntity.cs
@@ -1,5 +1,7 @@
-﻿using Hardware.Print.Tsc;
-using Hardware.Utils;
+﻿// This is an independent project of an individual developer. Dear PVS-Studio, please check it.
+// PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
+
+using Hardware.Print.Tsc;
 using log4net;
 using System;
 using System.Collections.Concurrent;
@@ -18,17 +20,14 @@ namespace Hardware.Print
         public int UserLabelCount { get; private set; }
         public PrinterStatus CurrentStatus { get; private set; }
         public int CommandThreadTimeOut { get; private set; } = 100;
-
         public delegate void OnHandler(PrintEntity state);
         public event OnHandler Notify;
-
         private readonly ILog _log = LogManager.GetLogger(System.Reflection.MethodBase.GetCurrentMethod().DeclaringType);
         public Connection Con { get; private set; }
         public ConcurrentQueue<string> CmdQueue { get; } = new ConcurrentQueue<string>();
         private readonly object _locker = new object();
         private Thread _sessionSharingThread = null;
         private bool _isThreadWork = true;
-
         public PrintControlEntity PrintControl { get; set; }
 
         #endregion
@@ -39,7 +38,7 @@ namespace Hardware.Print
         {
             CommandThreadTimeOut = commandThreadTimeOut;
             Con = connection;
-            PrintControl = new PrintControlEntity(Interface.Ethernet, ip, port);
+            PrintControl = new PrintControlEntity(PrintInterface.Ethernet, ip, port);
         }
 
         public PrintEntity(string ip, int port, int commandThreadTimeOut = 100)
@@ -47,7 +46,7 @@ namespace Hardware.Print
             //var zebraCurrentState = new StateEntity();
             CommandThreadTimeOut = commandThreadTimeOut;
             Con = new TcpConnection(ip, port);
-            PrintControl = new PrintControlEntity(Interface.Ethernet, ip, port);
+            PrintControl = new PrintControlEntity(PrintInterface.Ethernet, ip, port);
         }
 
         #endregion
@@ -139,11 +138,12 @@ namespace Hardware.Print
                                 request = request.Replace("|", "\\&");
                                 if (!request.Equals("^XA~JA^XZ") && !request.Contains("odometer.user_label_count"))
                                 {
-                                    //CurrentStatus = printerDevice.GetCurrentStatus();
-                                    //UserLabelCount = int.Parse(SGD.GET("odometer.user_label_count", printerDevice.Connection));
-                                    //UserLabelCount = 1;
-                                    //Peeler = SGD.GET("sensor.peeler", printerDevice.Connection);
-                                    PrintControl.SendCmd(false, request, false);
+                                    var taskPrint = new Task(async () =>
+                                    {
+                                        PrintControl.Cmd.SendCustom(false, request, false);
+                                        await Task.Delay(TimeSpan.FromMilliseconds(CommandThreadTimeOut)).ConfigureAwait(true);
+                                    });
+                                    taskPrint.Start();
                                 }
                             }
                             Notify?.Invoke(this);
@@ -191,7 +191,7 @@ namespace Hardware.Print
             }
             if (printerType.Contains("TSC "))
             {
-                PrintControl.ClearBuffer(false);
+                PrintControl.Cmd.ClearBuffer(true);
             }
             else
             {