+++ Hardware/Print/Zebra/DeviceContainer.cs
@@ -1,18 +1,21 @@
-﻿using System;
+﻿// This is an independent project of an individual developer. Dear PVS-Studio, please check it.
+// PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
+
+using System;
 using System.Collections.Generic;
 using System.Linq;
 
 namespace Hardware.Print.Zebra
 {
-    public class DeviceСontainer
+    public class DeviceContainer
     {
 
-        private static readonly Lazy<DeviceСontainer> _instance = new Lazy<DeviceСontainer>(() => new DeviceСontainer());
-        public static DeviceСontainer Instance => _instance.Value;
+        private static readonly Lazy<DeviceContainer> _instance = new Lazy<DeviceContainer>(() => new DeviceContainer());
+        public static DeviceContainer Instance => _instance.Value;
 
         private Dictionary<Guid, DeviceEntity> zebraDevices = null;
 
-        private DeviceСontainer()
+        private DeviceContainer()
         {
             zebraDevices = new Dictionary<Guid, DeviceEntity>();
         }